% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-annotation.R
\name{add_kmer}
\alias{add_kmer}
\title{Add centred k-mer sequence context}
\usage{
add_kmer(gr, genome, k = 5L, out_col = "kmer", seqstyle = NULL, chrs = NULL)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{genome}{A BSgenome or compatible genome object.}

\item{k}{K-mer width.}

\item{out_col}{Output metadata column.}

\item{seqstyle}{Optional sequence naming style.}

\item{chrs}{Optional seqlevels to keep.}
}
\value{
\code{gr} with \code{out_col} added.
}
\description{
Extracts a centred k-mer around each site from a genome object supported by
\code{Biostrings::getSeq()}. Windows that cannot be extracted at full width are set to NA.
}
\examples{
if (interactive()) {
    genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
    gr <- GenomicRanges::GRanges(
        "chr1", IRanges::IRanges(11874, width = 1), strand = "+"
    )
    add_kmer(gr, genome = genome, k = 5)
}

}
