% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-helpers.R
\name{makeID}
\alias{makeID}
\title{Add coordinate-derived IDs to a GRanges}
\usage{
makeID(gr, strand = TRUE, site_id_col = "site_id")
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{strand}{If TRUE, include strand in the ID.}

\item{site_id_col}{Metadata column to update. Set NULL to update names only.}
}
\value{
The input \code{GRanges} with names, and optionally \code{site_id_col}, set.
}
\description{
Add coordinate-derived IDs to a GRanges
}
\examples{
gr <- GenomicRanges::GRanges(
    "chr1", IRanges::IRanges(c(10, 20), width = 1), strand = "+"
)
makeID(gr)

}
