% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plots.R
\name{plot_distance_density}
\alias{plot_distance_density}
\title{Plot a generic distance density}
\usage{
plot_distance_density(
  gr,
  distance_col,
  set_col = NULL,
  positive_value = "positive",
  negative_value = "matched_negative",
  transform = c("log1p", "identity"),
  facet_by_location = TRUE,
  location_col = "location",
  bw_adjust = 1
)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{distance_col}{Numeric metadata column to plot.}

\item{set_col}{Optional grouping column.}

\item{positive_value}{Positive set value.}

\item{negative_value}{Matched-negative set value.}

\item{transform}{Either \code{"log1p"} or \code{"identity"}.}

\item{facet_by_location}{If TRUE, facet by \code{location_col}.}

\item{location_col}{Location column.}

\item{bw_adjust}{Density bandwidth multiplier.}
}
\value{
A ggplot object.
}
\description{
Plot a generic distance density
}
