% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plots.R
\name{plot_feature_width_bins}
\alias{plot_feature_width_bins}
\title{Plot feature-width quantile bins}
\usage{
plot_feature_width_bins(
  gr,
  set_col = NULL,
  positive_value = "positive",
  negative_value = "matched_negative",
  width_col = "feature_width",
  n_bins = 10,
  log1p_first = TRUE,
  facet_by_location = TRUE,
  location_col = "location",
  use_existing_bin_col = TRUE
)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{set_col}{Optional grouping column.}

\item{positive_value}{Positive value.}

\item{negative_value}{Negative value.}

\item{width_col}{Width column.}

\item{n_bins}{Number of quantile bins.}

\item{log1p_first}{If TRUE, bin log1p(width).}

\item{facet_by_location}{If TRUE, facet by location.}

\item{location_col}{Location column.}

\item{use_existing_bin_col}{If TRUE, use an existing matcher bin column named
\code{bin_<width_col>} when present; otherwise compute quantile bins from
\code{width_col}.}
}
\value{
A ggplot object.
}
\description{
Plot feature-width quantile bins
}
\examples{
gr <- GenomicRanges::GRanges(
    "chr1", IRanges::IRanges(seq(10, 100, by = 10), width = 1), strand = "+"
)
S4Vectors::mcols(gr)$feature_width <- seq(100, 1000, length.out = length(gr))
plot_feature_width_bins(gr, n_bins = 3, facet_by_location = FALSE)

}
