% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-background.R
\name{subset_matched_sets}
\alias{subset_matched_sets}
\title{Subset to matched positive/background rows}
\usage{
subset_matched_sets(gr, ...)
}
\arguments{
\item{gr}{A \code{GRanges} returned by \code{match_background()} or
\code{match_random_background()}.}

\item{...}{Passed to \code{\link{subset_matched_pairs}}.}
}
\value{
A paired \code{GRanges} containing only matched positives and their
selected matched negatives.
}
\description{
Convenience alias for \code{\link{subset_matched_pairs}}. This is the
object that should usually be passed to matched-set diagnostic plots.
}
\examples{
gr <- GenomicRanges::GRanges(
    "chr1", IRanges::IRanges(c(10, 20), width = 1), strand = "+"
)
gr <- prepare_sites(gr, label = c(1, 0))
S4Vectors::mcols(gr)$matched_negative_id <- c(names(gr)[2], NA)
S4Vectors::mcols(gr)$matched_positive_id <- c(NA, names(gr)[1])
subset_matched_sets(gr)

}
