% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-annotation.R
\name{add_feature_geometry}
\alias{add_feature_geometry}
\title{Compute within-feature geometry and metagene coordinates}
\usage{
add_feature_geometry(gr, resources)
}
\arguments{
\item{gr}{A \code{GRanges} with \code{location} and \code{tx_name} columns.}

\item{resources}{Output of \code{\link{build_tx_resources}}.}
}
\value{
The input \code{GRanges} with geometry columns added/replaced.
}
\description{
Adds feature proportions, equal-width metagene coordinates, median/mean split metagene
coordinates, transcript positions, distances to CDS start/stop, and distances to the
nearest exon junction where available.
}
\examples{
if (interactive()) {
    txdb <- get(
        "TxDb.Hsapiens.UCSC.hg38.knownGene",
        envir = asNamespace("TxDb.Hsapiens.UCSC.hg38.knownGene")
    )
    resources <- build_tx_resources(txdb)
    gr <- GenomicRanges::GRanges(
        "chr1", IRanges::IRanges(11874, width = 1), strand = "+"
    )
    gr <- annotate_sites(gr, txdb = txdb, resources = resources)
    add_feature_geometry(gr, resources = resources)
}

}
