% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-annotation.R
\name{add_metagene_split}
\alias{add_metagene_split}
\title{Add an isoform-averaged metagene split coordinate}
\usage{
add_metagene_split(gr, resources)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{resources}{Output of \code{\link{build_tx_resources}}.}
}
\value{
\code{gr} with \code{metagene_split} and \code{metagene_split3} replaced.
}
\description{
Expands each site to overlapping transcript isoforms, computes per-isoform metagene
coordinates, and collapses back to one mean coordinate per original site.
}
\examples{
if (interactive()) {
    txdb <- get(
        "TxDb.Hsapiens.UCSC.hg38.knownGene",
        envir = asNamespace("TxDb.Hsapiens.UCSC.hg38.knownGene")
    )
    resources <- build_tx_resources(txdb)
    gr <- GenomicRanges::GRanges(
        "chr1", IRanges::IRanges(11874, width = 1), strand = "+"
    )
    gr <- annotate_sites(gr, txdb = txdb, resources = resources)
    add_metagene_split(gr, resources = resources)
}

}
