% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-io.R
\name{as_basic_site_table}
\alias{as_basic_site_table}
\title{Return a small display table for annotated sites}
\usage{
as_basic_site_table(gr, compatibility_names = TRUE)
}
\arguments{
\item{gr}{An annotated \code{GRanges}.}

\item{compatibility_names}{Passed to \code{as_site_table()}.}
}
\value{
A compact data.frame.
}
\description{
Return a small display table for annotated sites
}
