% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-annotation.R
\name{as_site_table}
\alias{as_site_table}
\title{Convert annotated sites to a scalar data.frame}
\usage{
as_site_table(gr, compatibility_names = TRUE)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{compatibility_names}{If TRUE, include column names similar to the user's
earlier project wrapper, such as \code{primary_region_class}.}
}
\value{
A data.frame.
}
\description{
Convert annotated sites to a scalar data.frame
}
