% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-background.R
\name{assign_split_by_chromosome}
\alias{assign_split_by_chromosome}
\title{Assign train/test/validation split by chromosome}
\usage{
assign_split_by_chromosome(
  gr,
  split_col = "split",
  test_chrs = "chr9",
  val_chrs = "chr20",
  train_label = "training",
  test_label = "testing",
  val_label = "validation",
  normalize_chr_prefix = TRUE,
  warn_missing = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{split_col}{Output metadata column.}

\item{test_chrs}{Chromosomes assigned to testing.}

\item{val_chrs}{Chromosomes assigned to validation.}

\item{train_label}{Training label.}

\item{test_label}{Testing label.}

\item{val_label}{Validation label.}

\item{normalize_chr_prefix}{Try to harmonise \code{chr} prefix usage.}

\item{warn_missing}{Warn if requested chromosomes are absent.}

\item{overwrite}{If FALSE, error if \code{split_col} exists.}
}
\value{
\code{gr} with split column.
}
\description{
Assign train/test/validation split by chromosome
}
