% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_matching_functions.R
\name{makeID}
\alias{makeID}
\title{Add site id according to genomic coordinates to a GRange}
\usage{
makeID(gr, strand = TRUE)
}
\arguments{
\item{strand}{whether or not to include the strand as part of the id}
}
\value{
the same GRange with the names set
}
\description{
Add site id according to genomic coordinates to a GRange
}
