% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plots.R
\name{plot_junction_distance_density}
\alias{plot_junction_distance_density}
\title{Plot distance to nearest exon junction}
\usage{
plot_junction_distance_density(
  gr,
  distance_col = "nearest_exon_junction_dist",
  ...
)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{distance_col}{Junction distance column.}

\item{...}{Passed to \code{\link{plot_distance_density}}.}
}
\value{
A ggplot object.
}
\description{
Plot distance to nearest exon junction
}
\examples{
gr <- GenomicRanges::GRanges(
    "chr1", IRanges::IRanges(seq(10, 100, by = 10), width = 1), strand = "+"
)
S4Vectors::mcols(gr)$nearest_exon_junction_dist <- seq(10, 100, by = 10)
plot_junction_distance_density(gr, facet_by_location = FALSE)

}
