% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plots.R
\name{plot_kmer_counts}
\alias{plot_kmer_counts}
\title{Plot k-mer counts}
\usage{
plot_kmer_counts(
  gr,
  set_col = NULL,
  positive_value = "positive",
  negative_value = "matched_negative",
  kmer_col = "kmer",
  top_n = 20
)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{set_col}{Optional grouping column.}

\item{positive_value}{Positive value.}

\item{negative_value}{Negative value.}

\item{kmer_col}{K-mer column.}

\item{top_n}{Number of top k-mers to show.}
}
\value{
A ggplot object.
}
\description{
Plot k-mer counts
}
