% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-plots.R
\name{plot_splice_distance_density}
\alias{plot_splice_distance_density}
\title{Plot splice-proximity density}
\usage{
plot_splice_distance_density(
  gr,
  set_col = NULL,
  positive_value = "positive",
  negative_value = "matched_negative",
  start_col = "dist_from_feature_start",
  end_col = "dist_from_feature_end",
  transform = c("log1p", "identity"),
  facet_by_location = TRUE,
  location_col = "location",
  bw_adjust = 1
)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{set_col}{Optional grouping column.}

\item{positive_value}{Positive value.}

\item{negative_value}{Negative value.}

\item{start_col}{Distance-from-feature-start column.}

\item{end_col}{Distance-from-feature-end column.}

\item{transform}{Either \code{"log1p"} or \code{"identity"}.}

\item{facet_by_location}{If TRUE, facet by location.}

\item{location_col}{Location column.}

\item{bw_adjust}{Bandwidth multiplier.}
}
\value{
A ggplot object.
}
\description{
Uses \code{nearest_exon_junction_dist} when present. Otherwise it falls back to
\code{min(dist_from_feature_start, dist_from_feature_end)}.
}
\examples{
gr <- GenomicRanges::GRanges(
    "chr1", IRanges::IRanges(seq(10, 100, by = 10), width = 1), strand = "+"
)
S4Vectors::mcols(gr)$nearest_exon_junction_dist <- seq(10, 100, by = 10)
plot_splice_distance_density(gr, facet_by_location = FALSE)

}
