% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-background.R
\name{subset_bg_set}
\alias{subset_bg_set}
\title{Subset to positives and selected matched negatives}
\usage{
subset_bg_set(
  gr,
  set_col = "match_set",
  positive_value = "positive",
  negative_value = "matched_negative"
)
}
\arguments{
\item{gr}{A \code{GRanges}.}

\item{set_col}{Membership column.}

\item{positive_value}{Value marking positives.}

\item{negative_value}{Value marking matched negatives.}
}
\value{
A subset \code{GRanges}.
}
\description{
Subset to positives and selected matched negatives
}
